requirements.txt
requests
django
djangorestframework
djangorestframework-simplejwt
validators
psycopg2-binary
psycopg2